✓ Setting up command pattern layer and setup concrete commands for initial use cases. https://app.asana.com/0/0/1202159130237888/f Added Commands interface and also command manager that manages commands history and also manages all other operation required like trimming the list when new command is added between the list.
✓ setting up core data structure. https://app.asana.com/0/0/1202023384504324/f
Added ECS for this that saves all drawn objects as entities and all associated data as components. setting up any optimizations on how we store the objects should be easy as every thing is decoupled.

✓ Define a project that maintains the state of the scene. https://app.asana.com/0/0/1202023384504322/f
Project is set up to have single high level managers singalongs but the drawing data and rendering is defined per scene that allows us to possibly have multiple drawings open on same time also having all the data that we need to save/export associated to single scene.

✓ Create UI and define all control panels. https://app.asana.com/0/0/1202023384504320/f
UI is using ImGui and it is setup to have docking support so user can set the panels how ever it is required. and the initial setup is saved in the imgui.ini.

✓ Adding color palette UI and backend. https://app.asana.com/0/0/1202023384504333/f
Color palette is added as a new pannel uising IMgui and will be linked with tools manager exposing a common color when adding new elements and also loading the color from the color component when object is selected for editing.